+++ cnccode/business_classes/BUActivity.inc.php
@@ -350,11 +350,11 @@ class BUActivity extends Business
         $buCustomer = new BUCustomer($this);
         $buCustomer->getCustomerByID($customerID, $dsCustomer);
         $buSite = new BUSite($this);
-        $buSite->getSiteByID($customerID, $dsCustomer->getValue('delSiteNo'), $dsSite);
+        $buSite->getSiteByID($customerID, $dsCustomer->getValue(DBECustomer::DeliverSiteNo), $dsSite);
         $dsCallActivity->setUpdateModeInsert();
         $dsCallActivity->setValue('callActivityID', 0);
         $dsCallActivity->setValue('customerID', $customerID);
-        $dsCallActivity->setValue('siteNo', $dsCustomer->getValue('delSiteNo'));
+        $dsCallActivity->setValue('siteNo', $dsCustomer->getValue(DBECustomer::DeliverSiteNo));
         $dsCallActivity->setValue('contactID', $dsSite->getValue('invContactID'));
         $dsCallActivity->setValue('userID', $userID);
         $dsCallActivity->setValue('callActTypeID', '');
@@ -894,8 +894,8 @@ class BUActivity extends Business
             $buCustomer = new BUCustomer($this);
             $buCustomer->getCustomerByID($dbeProblem->getValue('customerID'), $dsCustomer);
             if (
-                $dsCustomer->getValue('specialAttentionFlag') == 'Y' &&
-                $dsCustomer->getValue('specialAttentionEndDate') >= date('Y-m-d')
+                $dsCustomer->getValue(DBECustomer::SpecialAttentionFlag) == 'Y' &&
+                $dsCustomer->getValue(DBECustomer::SpecialAttentionEndDate) >= date('Y-m-d')
             ) {
                 $this->sendSpecialAttentionEmail($dbeCallActivity->getPKValue());
             }
@@ -1269,7 +1269,7 @@ class BUActivity extends Business
     Send the email to all the main support email addresses at the client but exclude them if they were the reporting contact.
     */
         if (
-            $dsCustomer->getValue('othersEmailMainFlag') == 'Y' &&
+            $dsCustomer->getValue(DBECustomer::OthersEmailMainFlag) == 'Y' &&
             $mainSupportEmailAddresses = $buCustomer->getMainSupportEmailAddresses(
                 $dbeJCallActivity->getValue('customerID'),
                 $toEmail
@@ -3549,7 +3549,7 @@ is currently a balance of ';
         $buCustomer->getCustomerByID($customerID, $dsCustomer);
         $buSite = new BUSite($this);
 
-        $buSite->getSiteByID($customerID, $dsCustomer->getValue('delSiteNo'), $dsSite);
+        $buSite->getSiteByID($customerID, $dsCustomer->getValue(DBECustomer::DeliverSiteNo), $dsSite);
 
         // create new problem here
         $dbeProblem = new DBEProblem($this);
@@ -3658,7 +3658,8 @@ is currently a balance of ';
         $buCustomer = new BUCustomer($this);
         $buCustomer->getCustomerByID($_SESSION[$sessionKey]['customerID'], $dsCustomer);
 
-        if ($dsCustomer->getValue('specialAttentionFlag') == 'Y' && $dsCustomer->getValue('specialAttentionEndDate') >= date('Y-m-d')) {
+        if ($dsCustomer->getValue(DBECustomer::SpecialAttentionFlag) == 'Y' &&
+            $dsCustomer->getValue(DBECustomer::SpecialAttentionEndDate) >= date('Y-m-d')) {
             $this->sendSpecialAttentionEmail($dbeCallActivity->getPKValue());
         }
 
@@ -4306,8 +4307,8 @@ is currently a balance of ';
         $buCustomer = new BUCustomer($this);
         $buCustomer->getCustomerByID($dbeProblem->getValue('customerID'), $dsCustomer);
         if (
-            $dsCustomer->getValue('specialAttentionFlag') == 'Y' &&
-            $dsCustomer->getValue('specialAttentionEndDate') >= date('Y-m-d')
+            $dsCustomer->getValue(DBECustomer::SpecialAttentionFlag) == 'Y' &&
+            $dsCustomer->getValue(DBECustomer::SpecialAttentionEndDate) >= date('Y-m-d')
         ) {
             $this->sendSpecialAttentionEmail($dbeCallActivity->getPKValue());
         }
@@ -6246,7 +6247,7 @@ is currently a balance of ';
         $dbeCustomer = new DBECustomer($this);
         $dbeCustomer->getRow($dbeJProblem->getValue('customerID'));
 
-        if ($dbeCustomer->getValue('othersEmailMainFlag') == 'N') {
+        if ($dbeCustomer->getValue(DBECustomer::OthersEmailMainFlag) == 'N') {
 
             $copyEmailToMainContact = false;
 
@@ -6254,7 +6255,7 @@ is currently a balance of ';
 
             if (
                 $parameters['templateName'] == 'WorkCommencedEmail' &&
-                $dbeCustomer->getValue('workStartedEmailMainFlag') == 'N'
+                $dbeCustomer->getValue(DBECustomer::WorkStartedEmailMainFlag) == 'N'
             ) {
                 $copyEmailToMainContact = false;
             }
@@ -6400,13 +6401,33 @@ is currently a balance of ';
         $dbeCustomer = new DBECustomer($this);
         $dbeCustomer->getRow($customerID);
 
-        $slaHours = $dbeCustomer->getValue('slaP' . $priority);
+        $priorityValue = null;
+
+        switch ($priority) {
+            case 1:
+                $priorityValue = DBECustomer::SlaP1;
+                break;
+            case 2:
+                $priorityValue = DBECustomer::SlaP2;
+                break;
+            case 3:
+                $priorityValue = DBECustomer::SlaP3;
+                break;
+            case 4:
+                $priorityValue = DBECustomer::SlaP4;
+                break;
+            case 5:
+                $priorityValue = DBECustomer::SlaP5;
+                break;
+        }
+
+        $slaHours = $dbeCustomer->getValue($priorityValue);
         /*
     Special attention customers get half of normal SLA
     */
         if (
-            $dbeCustomer->getValue('specialAttentionFlag') == 'Y' &&
-            $dbeCustomer->getValue('specialAttentionEndDate') >= date('Y-m-d')
+            $dbeCustomer->getValue(DBECustomer::SpecialAttentionFlag) == 'Y' &&
+            $dbeCustomer->getValue(DBECustomer::SpecialAttentionEndDate) >= date('Y-m-d')
         ) {
             $slaHours = $slaHours / 2;
         }