+++ cnccode/controller_classes/CTCustomer.inc.php
@@ -605,7 +605,7 @@ class CTCustomer extends CTCNC
                     $_SERVER['PHP_SELF'],
                     array(
                         'action' => CTCNC_ACT_DISP_EDIT,
-                        'customerID' => $dsCustomer->getValue('CustomerID')
+                        'customerID' => $dsCustomer->getValue(DBECustomer::CustomerID)
                     )
                 );
             header('Location: ' . $nextURL);
@@ -645,12 +645,12 @@ class CTCustomer extends CTCNC
                         $_SERVER['PHP_SELF'],
                         array(
                             'action' => 'displayEditForm',
-                            'customerID' => $dsCustomer->getValue('CustomerID')
+                            'customerID' => $dsCustomer->getValue(DBECustomer::CustomerID)
                         )
                     );
 
-                if ($dsCustomer->getValue('reviewUserID')) {
-                    $buUser->getUserByID($dsCustomer->getValue('reviewUserID'), $dsUser);
+                if ($dsCustomer->getValue(DBECustomer::ReviewUserID)) {
+                    $buUser->getUserByID($dsCustomer->getValue(DBECustomer::ReviewUserID), $dsUser);
                     $user = $dsUser->getValue('name');
                 } else {
                     $user = false;
@@ -658,10 +658,10 @@ class CTCustomer extends CTCNC
 
                 $this->template->set_var(
                     array(
-                        'customerName' => $dsCustomer->getValue('Name'),
-                        'reviewDate' => $dsCustomer->getValue('reviewDate'),
-                        'reviewTime' => $dsCustomer->getValue('reviewTime'),
-                        'reviewAction' => $dsCustomer->getValue('reviewAction'),
+                        'customerName' => $dsCustomer->getValue(DBECustomer::Name),
+                        'reviewDate' => $dsCustomer->getValue(DBECustomer::ReviewDate),
+                        'reviewTime' => $dsCustomer->getValue(DBECustomer::ReviewTime),
+                        'reviewAction' => $dsCustomer->getValue(DBECustomer::ReviewAction),
                         'reviewUser' => $user,
                         'linkURL' => $linkURL
                     )
@@ -856,14 +856,14 @@ ORDER BY cus_name ASC  ";
                         $_SERVER['PHP_SELF'],
                         array(
                             'action' => 'dispEdit',
-                            'customerID' => $dsCustomer->getValue('CustomerID')
+                            'customerID' => $dsCustomer->getValue(DBECustomer::CustomerID)
                         )
                     );
 
 
                 $this->template->set_var(
                     array(
-                        'customerName' => $dsCustomer->getValue('Name'),
+                        'customerName' => $dsCustomer->getValue(DBECustomer::Name),
                         'linkURL' => $linkURL
                     )
                 );
@@ -910,15 +910,15 @@ ORDER BY cus_name ASC  ";
                         $_SERVER['PHP_SELF'],
                         array(
                             'action' => 'dispEdit',
-                            'customerID' => $dsCustomer->getValue('CustomerID')
+                            'customerID' => $dsCustomer->getValue(DBECustomer::CustomerID)
                         )
                     );
 
 
                 $this->template->set_var(
                     array(
-                        'customerName' => $dsCustomer->getValue('Name'),
-                        'specialAttentionEndDate' => $dsCustomer->getValue('specialAttentionEndDate'),
+                        'customerName' => $dsCustomer->getValue(DBECustomer::Name),
+                        'specialAttentionEndDate' => $dsCustomer->getValue(DBECustomer::SpecialAttentionEndDate),
                         'linkURL' => $linkURL
                     )
                 );