+++ cnccode/controller_classes/CTCustomerProfitabilityMonthsReport.inc.php
@@ -108,7 +108,7 @@ class CTCustomerProfitabilityMonthsReport extends CTCNC
         if ($dsSearchForm->getValue('customerID') != 0) {
             $buCustomer = new BUCustomer ($this);
             $buCustomer->getCustomerByID($dsSearchForm->getValue('customerID'), $dsCustomer);
-            $customerString = $dsCustomer->getValue('name');
+            $customerString = $dsCustomer->getValue(DBECustomer::Name);
         }
 
         $this->template->set_var(