+++ cnccode/controller_classes/CTCustomerProfitabilityReport.inc.php
@@ -110,7 +110,7 @@ class CTCustomerProfitabilityReport extends CTCNC
         if ($dsSearchForm->getValue('customerID') != 0) {
             $buCustomer = new BUCustomer ($this);
             $buCustomer->getCustomerByID($dsSearchForm->getValue('customerID'), $dsCustomer);
-            $customerString = $dsCustomer->getValue('name');
+            $customerString = $dsCustomer->getValue(DBECustomer::Name);
         }
 
         $totalCost = 0;