+++ cnccode/controller_classes/CTPassword.inc.php
@@ -182,7 +182,7 @@ class CTPassword extends CTCNC
                     'urlSubmit' => $urlSubmit,
                     'urlAdd' => $urlAdd,
                     'urlLoadFromCsv' => $urlLoadFromCsv,
-                    'customerName' => $dbeCustomer->getValue('Name'),
+                    'customerName' => $dbeCustomer->getValue(DBECustomer::Name),
                     'customerID' => $_REQUEST['customerID'],
                     'formError' => $this->getFormErrorMessage()
                 )